+++ cpukit/libnetworking/rtems/rtems_mii_ioctl.c
@@ -158,9 +158,9 @@ rtems_str2ifmedia (const char *str, int phy)
         return 0;
       if (!strncmp (chpt, "ase", 3))
         chpt += 3;
-      if (toupper (*chpt++) != 'T')
+      if (toupper ((int)*chpt++) != 'T')
         return 0;
-      if (IFM_100_TX == sub && toupper (*chpt++) != 'X')
+      if (IFM_100_TX == sub && toupper ((int)*chpt++) != 'X')
         return 0;
     }
   }
